adds all nav to remediationActivity
